requirements.txt
Flask
requests
python-firebase
pandas
sklearn
scikit-learn